Docker et Telemetry
- Need to copy settings.local.py to settings.py, and to adapt to the local changes (postgres password)
- need to pip3 install -r requirements.txt
- Documentation for telemtry will come later…
- no changes in the standard usage in local mode

+++ app/classes/calcul/calcObservation.py
@@ -9,6 +9,7 @@ import app.tools.myTools as t
 from app.tools.telemetry import Telemetry
 from django.db import transaction
 from django.conf import settings
+import threading
 import datetime
 import json
 import os
@@ -21,6 +22,7 @@ class CalcObs(AllCalculus):
 
     def __init__(self):
         self.tracer = Telemetry.Start("calculus")
+        CalcObs.lock = threading.Lock()
 
     def LoadJsonFromSvc(self, data: json):
         """
@@ -42,7 +44,6 @@ class CalcObs(AllCalculus):
         trace_flag = False
         if params.get("trace_flag") is not None:
             trace_flag = params["trace_flag"]
-        my_span = self.tracer.start_as_current_span("calculus", trace_flag)
 
         try:
             # load our parameters
@@ -62,9 +63,7 @@ class CalcObs(AllCalculus):
             # load the content of files on the server
             # getJsonData create a new span
             for j_data in self._getJsonData(params):
-                func_ret = self._loadJsonArrayInObs(
-                    j_data["j"], trace_flag, is_tmp, use_validation, j_data["f"]
-                )
+                func_ret = self.LoadJsonFromCall(j_data["j"], trace_flag, False, is_tmp, use_validation, j_data["f"])
                 ret.append(func_ret[0])
 
             # activate the computation of aggregations
@@ -72,7 +71,8 @@ class CalcObs(AllCalculus):
             return ret
 
         except Exception as inst:
-            my_span.record_exception(inst)
+            t.LogCritical(inst)
+            raise inst
 
     def LoadJsonFromCall(
         self,
@@ -83,13 +83,22 @@ class CalcObs(AllCalculus):
         use_validation: bool = False,
         filename: str = "???",
     ):
-        with self.tracer.start_as_current_span("calculus", trace_flag):
+        if CalcObs.lock.acquire(True, 500) is False:
+            t.logWarning("lock time-out !")
+            return
+        try:
+            self.tracer.add_future_attribute("json_file", filename)
+            self.tracer.save_future_attribute()
             # delete is not part of the transaction
             if delete_flag:
                 self.delete_obs_agg(is_tmp)
-            self._loadJsonArrayInObs(
-                json_data_array, trace_flag, is_tmp, use_validation, filename
-            )
+            ret = self._loadJsonArrayInObs(json_data_array, trace_flag, is_tmp, use_validation)
+            return ret
+        except Exception as inst:
+            t.LogCritical(inst)
+            raise inst
+        finally:
+            CalcObs.lock.release()
 
     # ----------------
     # private methods
@@ -101,7 +110,6 @@ class CalcObs(AllCalculus):
         trace_flag: bool = False,
         is_tmp: bool = None,
         use_validation: bool = False,
-        filename: str = "???",
     ) -> json:
         """
         processJson
@@ -117,30 +125,42 @@ class CalcObs(AllCalculus):
         if check_result is not None:
             raise Exception("Invalid Json: " + str(check_result))
 
-        my_span = self.tracer.get_current_or_new_span("calculus", trace_flag)
         idx = 0
-        while idx < json_data_array.__len__():
-            with self.tracer.start_span("Obs", trace_flag) as my_span:
-                if json_data_array.__len__() > 1:
+        try:
+            while idx < json_data_array.__len__():
+                # regenerate our future_attribute to all span(s) processing this json data
+                self.tracer.restore_future_attribute(False)
+                with self.tracer.start_as_current_span("Load Obs", trace_flag) as my_span:
+                    self.tracer.load_future_attributes_in_span(my_span)
                     my_span.set_attribute("idx", idx)
-                json_file_data = json_data_array[idx]
-                item_processed += json_file_data["data"].__len__()
-                ret = self._loadJsonItemInObs(
-                    json_file_data, trace_flag, is_tmp, use_validation
-                )
-                my_span.set_attribute(
-                    "items_processed", json_file_data["data"].__len__()
-                )
-                ret_data.append(ret)
-                idx += 1
+                    start_time = datetime.datetime.now()
+                    if json_data_array.__len__() > 1:
+                        my_span.set_attribute("idx", idx)
+                    json_file_data = json_data_array[idx]
+                    item_processed += json_file_data["data"].__len__()
+                    ret = self._loadJsonItemInObs(json_file_data, trace_flag, is_tmp, use_validation)
+                    my_span.set_attribute("items_processed", json_file_data["data"].__len__())
+                    duration = datetime.datetime.now() - start_time
+                    t.logInfo(
+                        "Json file loaded",
+                        my_span,
+                        {"time_exec": int(duration.microseconds / 1000), "items": item_processed, "idx": idx},
+                    )
+                    ret_data.append(ret)
+                    idx += 1
+        except Exception as exc:
+            t.LogCritical(exc)
+            raise exc
 
+        finally:
+            self.tracer.end_span()
+
+        global_duration = datetime.datetime.now() - debut_full_process
         ret_data.append(
             {
-                "total_exec": str(datetime.datetime.now() - debut_full_process),
+                "total_exec": int(global_duration.microseconds/1000),
                 "item_processed": item_processed,
-                "one_exec": str(
-                    (datetime.datetime.now() - debut_full_process) / item_processed
-                ),
+                "one_exec": int(global_duration.microseconds/1000 / item_processed),
             }
         )
         return ret_data
@@ -162,61 +182,40 @@ class CalcObs(AllCalculus):
 
         measure_idx = 0
         debut_process = datetime.datetime.now()
-        my_span = self.tracer.get_current_or_new_span("calculus", trace_flag)
+        my_span = self.tracer.get_current_or_new_span("Load Obs", trace_flag)
 
         while measure_idx < json_file_data["data"].__len__():
             # we use the stop_dat of our measure json as the start date for our processing
             m_stop_date_agg_start_date = json_file_data["data"][measure_idx]["stop_dat"]
 
-            if (
-                json_file_data["data"][measure_idx].get("current") is not None
-                and use_validation is False
-            ):
+            if (json_file_data["data"][measure_idx].get("current") is not None and use_validation is False):
                 m_duration = json_file_data["data"][measure_idx]["current"]["duration"]
             else:
                 # we don't care as we don't have current data
                 m_duration = 0
                 # in validation mode, we don't use the 'current'
                 json_file_data["data"][measure_idx]["current"] = {"duration": 0}
-            poste_metier = PosteMetier(
-                json_file_data["poste_id"], m_stop_date_agg_start_date
-            )
+            poste_metier = PosteMetier(json_file_data["poste_id"], m_stop_date_agg_start_date)
             if measure_idx == 0:
-                my_span.set_attribute("poste_id", str(poste_metier.data.id))
+                my_span.set_attribute("poste_id", poste_metier.data.id)
                 my_span.set_attribute("stopDat", str(m_stop_date_agg_start_date))
                 my_span.set_attribute("is_tmp", is_tmp)
             try:
                 poste_metier.lock()
-                obs_meteor = poste_metier.observation(
-                    m_stop_date_agg_start_date, is_tmp
-                )
-                if (
-                    obs_meteor.data.id is not None
-                    and json_file_data["data"][measure_idx].__contains__("update_me")
-                    is False
-                ):
+                obs_meteor = poste_metier.observation(m_stop_date_agg_start_date, is_tmp)
+                if (obs_meteor.data.id is not None and json_file_data["data"][measure_idx].__contains__("update_me") is False):
                     t.logInfo(
-                        "skipping data["
-                        + str(measure_idx)
-                        + "], stop_dat: "
-                        + str(m_stop_date_agg_start_date)
-                        + " already loaded",
+                        "skipping data[" + str(measure_idx) + "], stop_dat: " + str(m_stop_date_agg_start_date) + " already loaded",
                         my_span,
                     )
                     continue
                 # load aggregations data in obs_meteor.data.j_agg
                 m_agg_j = []
                 if use_validation is True:
-                    if json_file_data["data"][measure_idx].get('validation') is not None:
+                    if (json_file_data["data"][measure_idx].get("validation") is not None):
                         m_agg_j = json_file_data["data"][measure_idx]["validation"]
                     if m_agg_j.__len__() == 0:
-                        t.logInfo(
-                            "skipping data["
-                            + str(measure_idx)
-                            + "], no data in JSON !!! stop_dat: "
-                            + str(m_stop_date_agg_start_date),
-                            my_span,
-                        )
+                        t.logInfo("skipping data[" + str(measure_idx) + "], no data in JSON !!! stop_dat: " + str(m_stop_date_agg_start_date), my_span)
                         continue
                 else:
                     if json_file_data["data"][measure_idx].__contains__("aggregations"):
@@ -224,10 +223,7 @@ class CalcObs(AllCalculus):
                 obs_meteor.data.j_agg = m_agg_j
 
                 # load duration and stop_dat if not already loaded
-                if (
-                    obs_meteor.data.duration == 0
-                    and json_file_data["data"][measure_idx].get("current") is not None
-                ):
+                if (obs_meteor.data.duration == 0 and json_file_data["data"][measure_idx].get("current") is not None):
                     obs_meteor.data.duration = json_file_data["data"][measure_idx][
                         "current"
                     ]["duration"]
@@ -257,6 +253,8 @@ class CalcObs(AllCalculus):
 
                 # save our new data
                 obs_meteor.save()
+                my_span.set_attribute("obs_id_" + str(measure_idx), obs_meteor.data.id)
+
                 a_todo = AggTodoMeteor(obs_meteor.data.id, is_tmp)
                 a_todo.data.j_dv.append(delta_values)
                 if measure_idx < json_file_data["data"].__len__() <= 1:
@@ -267,28 +265,17 @@ class CalcObs(AllCalculus):
                 j_trace = {}
 
                 if trace_flag:
+                    duration2 = datetime.datetime.now() - debut_process
                     j_trace["info"] = "idx=" + str(measure_idx)
-                    j_trace["total_exec"] = str(datetime.datetime.now() - debut_process)
-                    j_trace["item_processed"] = str(json_file_data["data"].__len__())
-                    j_trace["one_exec"] = str(
-                        (datetime.datetime.now() - debut_process)
-                        / json_file_data["data"].__len__()
-                    )
+                    j_trace["total_exec"] = int(duration2.microseconds / 1000)
+                    j_trace["item_processed"] = json_file_data["data"].__len__()
+                    j_trace["one_exec"] = int(duration2.microseconds / 1000 / json_file_data["data"].__len__())
                     # j_trace['start_dat'] = json_file_data['data'][measure_idx]['current']['start_dat']
-                    j_trace["stop_dat"] = json_file_data["data"][measure_idx][
-                        "stop_dat"
-                    ]
-                    j_trace["obs data"] = JsonPlus().loads(
-                        JsonPlus().dumps(obs_meteor.data.j)
-                    )
-                    j_trace["obs aggregations"] = JsonPlus().loads(
-                        JsonPlus().dumps(obs_meteor.data.j_agg)
-                    )
-                    j_trace["agg_todo dv"] = (
-                        JsonPlus().loads(JsonPlus().dumps(a_todo.data.j_dv))
-                        if not (a_todo.data.id is None)
-                        else "{}"
-                    )
+                    j_trace["stop_dat"] = str(json_file_data["data"][measure_idx]["stop_dat"])
+                    j_trace["obs data"] = JsonPlus().loads(JsonPlus().dumps(obs_meteor.data.j))
+                    j_trace["obs aggregations"] = JsonPlus().loads(JsonPlus().dumps(obs_meteor.data.j_agg))
+                    j_trace["agg_todo dv"] = (JsonPlus().loads(JsonPlus().dumps(a_todo.data.j_dv)) if not (a_todo.data.id is None) else "{}")
+                    # t.LogDebug("json item loaded", my_span, {'idx': measure_idx, 'time_exec': j_trace["total_exec"], 'items': j_trace["item_processed"], "stop_dat": j_trace["stop_dat"]})
 
                 if j_trace != {}:
                     ret.append(j_trace)
@@ -301,10 +288,8 @@ class CalcObs(AllCalculus):
 
     def _getJsonData(self, params: json):
         """
-        yield filename, file_content
+            yield filename, file_content
         """
-        my_span = self.tracer.get_current_or_new_span("calculus", params["trace_flag"])
-
         try:
             # content loaded on client side
             if params.get("json") is not None:
@@ -313,11 +298,7 @@ class CalcObs(AllCalculus):
                 filename = "???"
                 if params.get("filename") is not None:
                     filename = params["filename"]
-                load_span = self.tracer.start_span("load", params["trace_flag"])
-                load_span.set_attribute("filename", filename)
                 yield {"f": filename, "j": my_json}
-                load_span.end()
-                t.logInfo("file loaded in Obs", None, {"filename": filename})
                 return
 
             # content to load from the server
@@ -325,10 +306,7 @@ class CalcObs(AllCalculus):
                 if hasattr(settings, "AUTOLOAD_DIR") is True:
                     params["base_dir"] = settings.AUTOLOAD_DIR
                 else:
-                    params["base_dir"] = (
-                        os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-                        + "/../../data/json_auto_load"
-                    )
+                    params["base_dir"] = (os.path.dirname(os.path.dirname(os.path.abspath(__file__))) + "/../../data/json_auto_load")
             base_dir = params["base_dir"]
             if params.get("filename") is not None:
                 files = [params["filename"]]
@@ -337,7 +315,7 @@ class CalcObs(AllCalculus):
 
             for aFile in files:
                 if aFile.endswith(".json"):
-                    load_span = self.tracer.start_span("load", params["trace_flag"])
+                    # with self.tracer.get_current_or_new_span("calculus", params["trace_flag"]) as load_span:
                     try:
                         # load our json file
                         texte = ""
@@ -347,30 +325,26 @@ class CalcObs(AllCalculus):
                                 texte += str(aligne)
                         my_json = JsonPlus().loads(texte)
 
-                        load_span.set_attribute("filename", aFile)
+                        # load_span.set_attribute("filename", aFile)
                         yield {"f": aFile, "j": my_json}
-                        load_span.add_event(
-                            "file.moved", {"dest": base_dir + "/done/" + aFile}
-                        )
-                        t.logInfo(
-                            "file processed, and moved",
-                            load_span,
-                            {"filename": aFile, "dest": base_dir + "/done/" + aFile},
-                        )
+                        # load_span.add_event("file.moved to [dest]", {"dest": base_dir + "/done/" + aFile})
                         os.rename(base_dir + "/" + aFile, base_dir + "/done/" + aFile)
+                        # t.logInfo(
+                        #     "json file loaded",
+                        #     load_span,
+                        #     {"filename": aFile, "dest": base_dir + "/done/" + aFile},
+                        # )
                     except Exception as exc:
-                        load_span.record_exception(exc)
-                        load_span.add_event(
-                            "file.moved", {"dest": base_dir + "/failed/" + aFile}
-                        )
+                        t.LogCritical(exc)
+                        # load_span.add_event("Exception, file moved to failed", {"dest": base_dir + "/failed/" + aFile})
                         t.LogError(
-                            "file NOT processed, and moved",
-                            load_span,
+                            "file moved to fail directory",
+                            None,
                             {"filename": aFile, "dest": base_dir + "/failed/" + aFile},
                         )
                         os.rename(base_dir + "/" + aFile, base_dir + "/failed/" + aFile)
-                    finally:
-                        load_span.end()
+                        raise exc
 
         except Exception as exc:
-            my_span.record_exception(exc)
+            t.LogCritical(exc)
+            raise exc
